+++ crates/eips/README.md
@@ -7,6 +7,7 @@ Contains constants, helpers, and basic data structures for consensus EIPs.
 ## Current support
 
 - EIP-1559
+- EIP-1898
 - EIP-2718
 - EIP-2930
 - EIP-4788
